+++ NLP/Dockerfile
@@ -41,6 +41,6 @@ COPY rest_api /home/user/rest_api
 
 EXPOSE 8000
 
-# cmd for running the API
+# Command for running the API server
 # gunicorn rest_api.application:app -k uvicorn.workers.UvicornWorker --workers 1
 CMD ["gunicorn", "rest_api.application:app",  "-b", "0.0.0.0", "-k", "uvicorn.workers.UvicornWorker", "--workers", "2", "--timeout", "180"]
