+++ NLP/requirements.txt
@@ -11,4 +11,5 @@ pytorch-lightning==1.3.8
 fastapi
 uvicorn
 gunicorn
+dill
 git+https://github.com/SKT-AI/KoBART.git#egg=kobart
