+++ NLP/riskout/fakenews/__init__.py
@@ -1,1 +1,2 @@
 from .fakenews_classifier import FakeNewsClassifier
+from .fakenews_classifier import get_tokenizer
