+++ NLP/riskout/fakenews/model/net.py
@@ -1,0 +1,33 @@
+import torch
+import torch.nn as nn
+
+from riskout.fakenews.model.utils import Vocab
+from riskout.fakenews.model.ops import MultiChannelEmbedding
+from riskout.fakenews.model.ops import ConvolutionLayer
+from riskout.fakenews.model.ops import MaxOverTimePooling
+
+
+class SenCNN(nn.Module):
+    """SenCNN class"""
+
+    def __init__(self, num_classes: int, vocab: Vocab) -> None:
+        """Instantiating SenCNN class
+        Args:
+            num_classes (int): the number of classes
+            vocab (model.utils.Vocab): the instance of model.utils.Vocab
+        """
+        super(SenCNN, self).__init__()
+        self._embedding = MultiChannelEmbedding(vocab)
+        self._convolution = ConvolutionLayer(300, 300)
+        self._pooling = MaxOverTimePooling()
+        self._dropout = nn.Dropout()
+        self._fc = nn.Linear(300, num_classes)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        fmap = self._embedding(x)
+        fmap = self._convolution(fmap)
+        feature = self._pooling(fmap)
+        feature = self._dropout(feature)
+        score = self._fc(feature)
+
+        return score
